+++ apps/extension/entrypoints/style.css
@@ -17,6 +17,7 @@
   --color-matter-05: #524e4b;
   --color-matter-06: #6d6965;
   --color-error: #942922;
+  --color-critical: #0C27D8;
   --color-red-10: #ed4136;
   --color-quantum-40: rgba(255, 71, 0, 0.4);
   --color-quantum-10: rgba(255, 71, 0, 0.1);
