+++ apps/extension/src/routes/send-token.tsx
@@ -3,10 +3,13 @@ import {
   SendTokenScreen,
   useAuthStore,
 } from '@evevault/shared'
+import { getHeaderIdentity } from '@evevault/shared/auth'
+import { SubpageHeader } from '@evevault/shared/components'
 import type { SendTokenSearch } from '@evevault/shared/router'
 import { localnetKeyService } from '@evevault/shared/services/keeperService'
 import { useContextStore } from '@evevault/shared/stores'
 import { EXTENSION_ROUTES } from '@evevault/shared/utils'
+import { useActiveSuiAddress } from '@evevault/shared/wallet'
 import {
   createFileRoute,
   redirect,
@@ -17,16 +20,24 @@ import {
 function SendTokenPage() {
   const navigate = useNavigate()
   const { user } = useAuthStore()
+  const activeAddress = useActiveSuiAddress()
   const { coinType } = useSearch({ from: '/send-token' })
 
   const handleNavigateBack = () => {
     navigate({ to: '/' })
   }
 
-  // SendTokenScreen renders its own SubpageHeader; user is optional because
-  // localnet has no zkLogin user (the header falls back to the active address).
+  // On localnet there is no zkLogin user; fall back to the active address.
+  const identity = user ? getHeaderIdentity(user) : { email: '', address: '' }
+
   return (
     <div className="flex flex-col gap-10">
+      <SubpageHeader
+        title="Transfer token"
+        email={identity.email}
+        address={activeAddress ?? identity.address}
+        onBack={handleNavigateBack}
+      />
       <SendTokenScreen
         coinType={coinType}
         user={user}
